+++ ArduinoLibs/SkaarhojUniSketch/SK_DEV_VIDEOHUB.h
@@ -72,7 +72,7 @@ uint16_t VideoHub_pullFromHoldGroup(uint8_t hgIdx, uint8_t HWc) {
 // 5 = dimmed
 // 1,2,3,4 = full (yellow), red, green, yellow
 // Bit 4 (16) = blink flag, filter out for KP01 buttons.
-uint16_t evaluateAction_VIDEOHUB(const uint8_t devIndex, const uint16_t actionPtr, const uint8_t HWc, const uint8_t actIdx, bool actDown, bool actUp, int pulses, int value) {
+uint16_t evaluateAction_VIDEOHUB(const uint8_t devIndex, const uint16_t actionPtr, const uint8_t HWc, const uint8_t actIdx, bool actDown, bool actUp, int16_t pulses, int16_t value) {
   uint16_t retVal = 0;
   if (actDown || actUp) {
     Serial << "VIDEO HUB action " << globalConfigMem[actionPtr] << "\n";
